Added two new functions to address this challenge.
The first function uses groupby (not that I'm importing the function from the itertools module at the beginning of the file). Must say that this is my favorite approach,
The second function uses a dictionary to keep the scores as its keys, and lists of the names as its values.